fix: quest realtime thing
