+++ packages/shared/src/components/quest/QuestButton.spec.tsx
@@ -5,21 +5,12 @@ import { act, render, screen, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import type { ReactElement, ReactNode } from 'react';
 import { TestBootProvider } from '../../../__tests__/helpers/boot';
-import type { QuestDashboard } from '../../graphql/quests';
-import {
-  QuestRewardType,
-  QuestStatus,
-  QuestType,
-  QUEST_ROTATION_UPDATE_SUBSCRIPTION,
-  QUEST_UPDATE_SUBSCRIPTION,
-} from '../../graphql/quests';
+import { QuestRewardType, QuestStatus, QuestType } from '../../graphql/quests';
 import { QuestButton } from './QuestButton';
 import { useClaimQuestReward } from '../../hooks/useClaimQuestReward';
 import { useQuestDashboard } from '../../hooks/useQuestDashboard';
 import { usePlusSubscription } from '../../hooks/usePlusSubscription';
-import useSubscription from '../../hooks/useSubscription';
 import { LogEvent, TargetId, TargetType } from '../../lib/log';
-import { generateQueryKey, RequestKey } from '../../lib/query';
 
 function mockReactModule() {
   return React;
@@ -49,11 +40,6 @@ jest.mock('../icons', () => {
     ),
   };
 });
-
-jest.mock('../../hooks/useSubscription', () => ({
-  __esModule: true,
-  default: jest.fn(),
-}));
 
 jest.mock('../../hooks/usePlusSubscription', () => ({
   usePlusSubscription: jest.fn(),
@@ -178,7 +164,6 @@ jest.mock('../tooltip/Tooltip', () => {
 
 const mockUseQuestDashboard = useQuestDashboard as jest.Mock;
 const mockUseClaimQuestReward = useClaimQuestReward as jest.Mock;
-const mockUseSubscription = useSubscription as jest.Mock;
 const mockUsePlusSubscription = usePlusSubscription as jest.Mock;
 const mockLogSubscriptionEvent = jest.fn();
 const mockPush = jest.fn();
@@ -261,7 +246,6 @@ beforeEach(() => {
     isPending: false,
     variables: undefined,
   });
-  mockUseSubscription.mockReset();
   mockLogSubscriptionEvent.mockReset();
   mockUsePlusSubscription.mockReturnValue({
     isPlus: false,
@@ -613,189 +597,6 @@ describe('QuestButton', () => {
     expect(
       screen.queryByTestId('quest-button-new-indicator'),
     ).not.toBeInTheDocument();
-  });
-
-  it('should log when a quest becomes claimable after a quest update', () => {
-    const logEvent = jest.fn();
-    const client = new QueryClient();
-    const subscriptions: Array<{
-      query: string;
-      next?: () => unknown;
-    }> = [];
-    let questDashboardState: {
-      data: QuestDashboard;
-      isPending: boolean;
-      isError: boolean;
-      dataUpdatedAt: number;
-    } = {
-      data: questDashboard,
-      isPending: false,
-      isError: false,
-      dataUpdatedAt: 1,
-    };
-
-    mockUseQuestDashboard.mockImplementation(() => questDashboardState);
-    mockUseSubscription.mockImplementation(
-      (
-        request: () => { query: string },
-        callbacks: { next?: () => unknown },
-      ) => {
-        subscriptions.push({
-          query: request().query,
-          next: callbacks.next,
-        });
-      },
-    );
-
-    const view = render(
-      <TestBootProvider client={client} log={{ logEvent }}>
-        <QuestButton />
-      </TestBootProvider>,
-    );
-
-    questDashboardState = {
-      data: {
-        ...questDashboard,
-        daily: {
-          ...questDashboard.daily,
-          regular: [
-            {
-              ...questDashboard.daily.regular[0],
-              userQuestId: 'user-quest-1',
-              status: QuestStatus.Completed,
-              claimable: true,
-              completedAt: new Date('2026-04-13T10:00:00.000Z'),
-            },
-          ],
-        },
-      },
-      isPending: false,
-      isError: false,
-      dataUpdatedAt: 2,
-    };
-
-    subscriptions
-      .find((subscription) => subscription.query === QUEST_UPDATE_SUBSCRIPTION)
-      ?.next?.();
-
-    view.rerender(
-      <TestBootProvider client={client} log={{ logEvent }}>
-        <QuestButton />
-      </TestBootProvider>,
-    );
-
-    expect(logEvent).toHaveBeenCalledWith({
-      event_name: LogEvent.QuestClaimable,
-      target_id: 'quest-1',
-      target_type: TargetType.Quest,
-      extra: JSON.stringify({
-        questType: QuestType.Daily,
-        userQuestId: 'user-quest-1',
-        userId: undefined,
-        rotationId: 'daily-quest-1',
-      }),
-    });
-  });
-
-  it('should not log claimable for a locked plus quest after a quest update', () => {
-    const logEvent = jest.fn();
-    const client = new QueryClient();
-    const subscriptions: Array<{
-      query: string;
-      next?: () => unknown;
-    }> = [];
-    let questDashboardState: {
-      data: QuestDashboard;
-      isPending: boolean;
-      isError: boolean;
-      dataUpdatedAt: number;
-    } = {
-      data: {
-        ...questDashboard,
-        daily: {
-          ...questDashboard.daily,
-          plus: [
-            {
-              rotationId: 'daily-plus-quest-1',
-              userQuestId: null,
-              progress: 1,
-              status: QuestStatus.InProgress,
-              locked: true,
-              claimable: false,
-              quest: {
-                id: 'plus-quest-1',
-                name: 'Plus read posts',
-                description: 'Read 3 posts today',
-                type: QuestType.Daily,
-                eventType: 'read_post',
-                targetCount: 3,
-              },
-              rewards: [{ type: QuestRewardType.Xp, amount: 150 }],
-            },
-          ],
-        },
-      },
-      isPending: false,
-      isError: false,
-      dataUpdatedAt: 1,
-    };
-
-    mockUseQuestDashboard.mockImplementation(() => questDashboardState);
-    mockUseSubscription.mockImplementation(
-      (
-        request: () => { query: string },
-        callbacks: { next?: () => unknown },
-      ) => {
-        subscriptions.push({
-          query: request().query,
-          next: callbacks.next,
-        });
-      },
-    );
-
-    const view = render(
-      <TestBootProvider client={client} log={{ logEvent }}>
-        <QuestButton />
-      </TestBootProvider>,
-    );
-
-    questDashboardState = {
-      data: {
-        ...questDashboardState.data,
-        daily: {
-          ...questDashboardState.data.daily,
-          plus: [
-            {
-              ...questDashboardState.data.daily.plus[0],
-              userQuestId: 'user-plus-quest-1',
-              status: QuestStatus.Completed,
-              progress: 3,
-              completedAt: new Date('2026-04-13T10:00:00.000Z'),
-            },
-          ],
-        },
-      },
-      isPending: false,
-      isError: false,
-      dataUpdatedAt: 2,
-    };
-
-    subscriptions
-      .find((subscription) => subscription.query === QUEST_UPDATE_SUBSCRIPTION)
-      ?.next?.();
-
-    view.rerender(
-      <TestBootProvider client={client} log={{ logEvent }}>
-        <QuestButton />
-      </TestBootProvider>,
-    );
-
-    expect(logEvent).not.toHaveBeenCalledWith(
-      expect.objectContaining({
-        event_name: LogEvent.QuestClaimable,
-        target_id: 'plus-quest-1',
-      }),
-    );
   });
 
   it('should stay open when the page scrolls', async () => {
@@ -1326,57 +1127,6 @@ describe('QuestButton', () => {
     expect(mockLogSubscriptionEvent).toHaveBeenCalledWith({
       event_name: LogEvent.UpgradeSubscription,
       target_id: TargetId.QuestDropdown,
-    });
-  });
-
-  it('should invalidate the quest dashboard on quest progress and rollover updates', () => {
-    const subscriptions: Array<{
-      query: string;
-      next?: () => unknown;
-    }> = [];
-    const client = new QueryClient();
-    const invalidateQueries = jest
-      .spyOn(client, 'invalidateQueries')
-      .mockResolvedValue(undefined);
-
-    mockUseSubscription.mockImplementation(
-      (
-        request: () => { query: string },
-        callbacks: { next?: () => unknown },
-      ) => {
-        subscriptions.push({
-          query: request().query,
-          next: callbacks.next,
-        });
-      },
-    );
-
-    renderComponent({ client });
-
-    expect(
-      Array.from(
-        new Set(subscriptions.map((subscription) => subscription.query)),
-      ),
-    ).toEqual([QUEST_UPDATE_SUBSCRIPTION, QUEST_ROTATION_UPDATE_SUBSCRIPTION]);
-
-    subscriptions
-      .find((subscription) => subscription.query === QUEST_UPDATE_SUBSCRIPTION)
-      ?.next?.();
-    subscriptions
-      .find(
-        (subscription) =>
-          subscription.query === QUEST_ROTATION_UPDATE_SUBSCRIPTION,
-      )
-      ?.next?.();
-
-    expect(invalidateQueries).toHaveBeenCalledTimes(2);
-    expect(invalidateQueries).toHaveBeenNthCalledWith(1, {
-      queryKey: generateQueryKey(RequestKey.QuestDashboard),
-      exact: true,
-    });
-    expect(invalidateQueries).toHaveBeenNthCalledWith(2, {
-      queryKey: generateQueryKey(RequestKey.QuestDashboard),
-      exact: true,
     });
   });
 });
